git-send-email: Use sanitized address when reading mbox body
Addresses that are mentioned on the trailers in the commit log
messages (e.g., "Reviewed-by") are added to the "Cc:" list by "git
send-email".  These hand-written addresses may be malformed (e.g.,
having unquoted "." and other punctutation marks in the display-name
part).

The code does use sanitize_address on these address-looking strings
to make them into valid addresses, but it is used only to see if the
address should be suppressed.  The original string taken from the
message is added to the @cc list.

Because the addresses on trailer lines are hand-written and more
likely to contain malformed addresses, when adding to the @cc list,
use the result from sanitize_address, not the original.

Note that we do not modify the behaviour for addresses taken from
the e-mail headers, as they are more likely to be machine generated
and well-formed.

Signed-off-by: Csókás, Bence <csokas.bence@prolan.hu>
Signed-off-by: Junio C Hamano <gitster@pobox.com>